current_state of new objects now returns the initial state correctly
When an object is freshly initialized, the current_State virtual
attribute calls on read_state to get an object's current state, which in
turn calls to_sym on the `state` attribute. However, the state attribute
of a newly initialized object might return nil, and this causes an
exception to be thrown.