Added tests for calling the failure and success callbacks
This uncovered some bug not firering the event_failed callback
in any case of an exception has been raised. This is now handled.

The exception is just re-raised if the record doesn't have defined
a event_failed callback. If he had it is up to the implementation
to handle the execption. We are just return false.